+++ packages/freezed/CHANGELOG.md
@@ -3,6 +3,8 @@
 - The generated `copyWith` is now annotated by `@useResult` (thanks to @miDeb)
 - Improved performance of `copyWith` (thanks to @miDeb)
 - Improved type inference when using `mapOrNull`/`whenOrNull` (thanks to @DevNico) 
+- Re-introduced `@With.fromString` and `@Implements.fromString` to allow unions
+  to implement generic types. (thanks to @rorystephenson)
 - fixes @Default for Strings containing (, [ or { (thanks to @hugobrancowb)
 - fix Freezed incorrectly comparing primitives using `DeepCollectionEquality` (thanks to @knaeckeKami)
 - fix image links (thanks to @SunlightBro)